moved bug finder logic into own class
