+++ src/main/java/at/hannibal2/skyhanni/mixins/transformers/MixinKeyBinding.java
@@ -2,7 +2,7 @@ package at.hannibal2.skyhanni.mixins.transformers;
 
 import at.hannibal2.skyhanni.data.model.TextInput;
 import at.hannibal2.skyhanni.features.garden.farming.GardenCustomKeybinds;
-import at.hannibal2.skyhanni.test.GraphEditor;
+import at.hannibal2.skyhanni.test.graph.GraphEditor;
 import net.minecraft.client.settings.KeyBinding;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
